Move integration-like tests into submodule
